+++ src/data/PostDatabase.ts
@@ -33,6 +33,7 @@ export class PostDatabase extends BaseDatabase {
 			.select("*")
 			.from(PostDatabase.TABLE_NAME)
 			.where({ id });
+			console.log("result getById: ", result)
         return result[0];
     }
